requirements.txt
APScheduler==3.6.3
CacheControl==0.12.11
cachetools==4.2.2
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
idna==3.4
msgpack==1.0.4
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rsa==4.9
six==1.16.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
